feat: miscellaneous
Mostly fixing random bugs or adding random stuff. NOTE TO SELF: There is a bug with the API when editing the current user. Somehow when modifying display_name, it doesn't change. HMM?

Also, TODO is to fork GenRegistry and make it watch processes to automatically remove them from the registry.